+++ frontend/compas_threejs/src/components/layout/Openbar.vue
@@ -1,5 +1,5 @@
 <template>
-    <div id="openbar" class="theme" :class="{ 'is-hidden': !isVisible }">
+    <div id="openbar" class="fixed-openbar theme" :class="{ 'is-hidden': !isVisible }">
         <!-- Dynamically render components from the store -->
         <div
             v-for="item in sidebarComponents"
@@ -7,8 +7,8 @@
             class="dynamic-item"
         >
             <!-- Add a label if it exists -->
-            <label v-if="item.label" class="dynamic-label">{{
-                item.label
+            <label v-if="item.label" class="dynamic-label" :class="{ dark: theme.value === 'dark' }">{{
+            item.label
             }}</label>
 
             <!-- Render a Button -->
@@ -77,23 +77,22 @@
             <ArrowBigLeftDash />
         </Button>
     </div>
-
-    <Button
-        variant="secondary"
-        size="icon"
-        class="mb-5"
-        @click="toggleSideBar()"
-        :class="{ 'is-hidden': !isVisible }"
-    >
-        <ArrowBigRightDash />
-    </Button>
+        <Button
+            variant="secondary"
+            size="icon"
+            class="mb-5"
+            @click="toggleSideBar()"
+            :class="{ 'is-hidden': !isVisible }"
+        >
+            <ArrowBigRightDash />
+        </Button>
 </template>
 
 <script setup lang="ts">
-import { watchEffect, toRaw } from "vue"; // 1. Import watchEffect and toRaw
+import { ref } from "vue";
+import { useEventListener } from "@vueuse/core";
 import { Button } from "@/components/ui/button";
 import { Slider } from "@/components/ui/slider";
-import { ref } from "vue"; // Import ref
 import { sidebarComponents, handleAction } from "@/communications/sidebarStore";
 import { ArrowBigLeftDash, ArrowBigRightDash } from "lucide-vue-next";
 import {
@@ -103,6 +102,7 @@ import {
     NumberFieldIncrement,
     NumberFieldInput,
 } from "@/components/ui/number-field";
+import { theme } from "@/store/store";
 
 const isVisible = ref(true);
 
@@ -161,9 +161,14 @@ div#openbar.is-hidden {
 .dynamic-label {
     font-weight: 500;
     font-size: 15px;
-    color: #333;
+    color: var(--foreground);
     padding: 0;
 }
+
+.slider-value {
+    color: var(--foreground);
+}
+
 Button.mb-4 {
     position: relative;
     margin: 0px;