+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Added `Toolbar.vue` featuring existing transforms and views shortcuts into user-friendly buttons.
 * Added `NumberField` class to `compas_threejs.ui`.
 * Added commands to save views, select them and download a screenshot as image.
+* Added dark/light theme toggle.
 * Added command to play/pause object motion.
 
 ### Changed